requirements.txt
awscli-local==0.20
psycopg2-binary==2.9.5
numpy==1.24.1
pandas==1.5.2
murmurhash3==2.3.5